converted Badge.stories.js to .tsx and updated stories to new CSF
